rudimentary tests for CRUD

+++ tests.py
@@ -2,6 +2,11 @@
 
 from pyfc4.models import *
 
+# logging
+import logging
+logging.basicConfig(level=logging.DEBUG)
+logger = logging.getLogger(__name__)
+
 # instantiate repository
 repo = Repository('http://localhost:8080/rest','ghukill','password', context={'foo':'http://foo.com'})
 
@@ -19,6 +24,7 @@ def get_bc():
 # create foo/bar (basic container)
 def create_child_bc():
 	bar = BasicContainer(repo, 'foo/bar')
+	bar.create()
 	return bar.exists()
 
 # get foo/bar from foo.children()
@@ -28,24 +34,33 @@ def get_child_bc():
 
 # create foo/baz (NonRDF / binary), from foo
 def create_child_binary():
-	baz = BasicContainer(repo, 'foo/bar')
+	baz = Binary(repo, 'foo/baz')
+	baz.data = 'this is a test, this is only a test'
+	baz.headers['Content-Type'] = 'text/plain'
+	baz.create()
+	return baz.exists()
 
 # get foo/baz
 def get_child_binary():
-	pass
+	baz = repo.get_resource('foo/baz')
+	return baz.exists()
 
 # delete all three
 def delete_test_resources():
 	for uri in ['foo','foo/bar','foo/baz']:
-		resource = repo.get_resource(uri)
-		resource.delete(remove_tombstone=True)
+		try:
+			resource = repo.get_resource(uri)
+			resource.delete(remove_tombstone=True)
+		except:
+			logger.debug('could not delete %s' % uri)
 
-
-def run_all_tests():
+# run all tests
+def run_all_tests(cleanup=False):
 	create_bc()
 	get_bc()
 	create_child_bc()
 	get_child_bc()
 	create_child_binary()
 	get_child_binary()
-	delete_test_resources()
+	if cleanup:
+		delete_test_resources()
